Revise error check
