+++ Cargo.toml
@@ -21,3 +21,6 @@ dirs = "1.0"
 failure = "0.1"
 log = "0.4"
 structopt = "0.2"
+
+[package.metadata.nvptx]
+runtime = ["core"]
